requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose
passlib
bcrypt
python-multipart
python-dotenv
email-validator

